pyproject.toml
[project]
name = "trabalho02-sistemas-distribuidos"
version = "0.1.0"
description = ""
authors = [
    {name = "armandoluz",email = "armandoari.288@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "kivy (>=2.3.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "plyer (>=2.1.0,<3.0.0)",
    "buildozer (>=1.5.0,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "setuptools (>=76.1.0,<77.0.0)",
    "cython (>=3.0.12,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
